This Data Science project is based on Covid-19 WHO Dataset, downloaded from Kaggle. Data is available for the period till 27th July, 2020 starting from 22nd January, 2020. Data is available for all the countries in the World. Goal is to visualize and analyze the data for selected countries, all the continents/regions and also for the entire world. In this study, pandas library in Python is used for various Dataframe operations (to make the data structured) and plotly (express, graph_objectsis etc.) is used for interactive visualizations (various interactive plots, Choropleth map etc.). We could extract various meaningful and interesting insight from the data visualizations and exploratory data analysis.
This Data Science project is based on Covid-19 WHO Dataset, downloaded from Kaggle. Data is available for the period till 27th July, 2020 starting from 22nd January, 2020. Data is available for all the countries in the World. Goal is to visualize and analyze the data for selected countries, all the continents/regions and also for the entire world. In this study, pandas library in Python is used for various Dataframe operations (to make the data structured) and plotly (express, graph_objectsis etc.) is used for interactive visualizations (various interactive plots, Choropleth map etc.). We could extract various meaningful and interesting insight from the data visualizations and exploratory data analysis.